chore(input): rollup svg plugin, genertateId update, moved svg to shared
