+++ lerna.json
@@ -1,7 +1,5 @@
 {
   "npmClient": "npm",
-  "packages": [
-    "packages/*"
-  ],
+  "packages": ["packages/*", "shared/**"],
   "version": "independent"
 }
