+++ packages/Input/Input.stories.js
@@ -24,7 +24,7 @@ export const Playground = () => (
     label={text('Label', 'Label')}
     placeholder={text('Placeholder', 'Placeholder')}
     error={text('Error Message', 'Error')}
-    feedback={select('Feedback', ['success', 'error'])}
+    feedback={select('Feedback', [undefined, 'success', 'error'])}
     small={boolean('Small Label', false)}
     feedbackicon={boolean('Feedback Icon', false)}
     required={boolean('Required Input', false)}
