+++ shared/utils/generateId/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "@pm-kit/util-generate-id",
+  "version": "1.0.0",
+  "private": true,
+  "main": "",
+  "module": "",
+  "dependencies": {
+    "array-find-es6": "^2.0.3"
+  }
+}
